Exception handling

+++ dscommerce/src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -1,38 +1,42 @@
 package com.devsuperior.dscommerce.services;
 
-import java.util.List;
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.dto.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repository.ProductRepository;
+import com.devsuperior.dscommerce.services.exeptions.DatabaseExeption;
+import com.devsuperior.dscommerce.services.exeptions.ResourceNotFoundExeption;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
-	
+
 	@Autowired
 	private ProductRepository repository;
-	
+
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Optional<Product> result = repository.findById(id);
-		Product product = result.get();
+		Product product = repository.findById(id)
+				.orElseThrow(() -> new ResourceNotFoundExeption("Recurso não encontrado"));
 		ProductDTO dto = new ProductDTO(product);
 		return dto;
 	}
-	
+
 	@Transactional(readOnly = true)
 	public Page<ProductDTO> findAll(Pageable pageable) {
 		Page<Product> product = repository.findAll(pageable);
 		return product.map(x -> new ProductDTO(x));
 	}
-	
+
 	@Transactional
 	public ProductDTO insert(ProductDTO dto) {
 		Product entity = new Product();
@@ -40,17 +44,32 @@ public class ProductService {
 		entity = repository.save(entity);
 		return new ProductDTO(entity);
 	}
-	
+
 	@Transactional
-	public ProductDTO update(Long id,ProductDTO dto) {
-		Product entity = repository.getReferenceById(id);
-		copyDtoToEntity(dto, entity);
-		return new ProductDTO(entity);
+	public ProductDTO update(Long id, ProductDTO dto) {
+		try {
+			Product entity = repository.getReferenceById(id);
+			copyDtoToEntity(dto, entity);
+			return new ProductDTO(entity);
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundExeption("Recurso não encontrado");
+		}
 	}
-	
-	@Transactional()
+
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundExeption("Recurso não encontrado");
+		}
+		try {
 		repository.deleteById(id);
+		}
+		catch(EmptyResultDataAccessException e){
+			throw new ResourceNotFoundExeption("Recurso não encontrado");
+		}
+		catch(DataIntegrityViolationException e) {
+			throw new DatabaseExeption("Falha de integridade referencial");
+		}
 	}
 
 	private void copyDtoToEntity(ProductDTO dto, Product entity) {
